fix type errors, set up multi proj

+++ py/dagster/curate1/resources/database/database_resource.py
@@ -1,13 +1,13 @@
 from abc import ABC, abstractmethod
 from datetime import datetime
-from typing import List, Optional
+from typing import Any, List, Optional
 
 from dagster import ConfigurableResource, InitResourceContext
 
 from .database import Database, Document, DocumentAttribute
 
 
-class DatabaseResource(ConfigurableResource, ABC):
+class DatabaseResource(ConfigurableResource[Any], ABC): # type: ignore
     @abstractmethod
     def insert_documents(self, documents: List[Document]) -> List[int]:
         pass
@@ -27,20 +27,28 @@ class DatabaseResource(ConfigurableResource, ABC):
 
 class SqliteDatabaseResource(DatabaseResource):
     db_path: str
-    _database: Database
+    _database: Optional[Database] = None
 
     def setup_for_execution(self, context: InitResourceContext) -> None:
         self._database = Database(context.resource_config["db_path"])
 
-    def insert_documents(self, documents: List[Document]) -> None:
+    def insert_documents(self, documents: List[Document]) -> List[int]:
+        if self._database is None:
+            raise ValueError("Database is not initialized.")
         return self._database.insert_documents(documents)
 
-    def insert_document_attributes(self, document_attributes: List[DocumentAttribute]) -> None:
+    def insert_document_attributes(self, document_attributes: List[DocumentAttribute]) -> List[int]:
+        if self._database is None:
+            raise ValueError("Database is not initialized.")
         return self._database.insert_document_attributes(document_attributes)
     
     def delete_documents_partition(self, partition_start: datetime, partition_end: datetime):
+        if self._database is None:
+            raise ValueError("Database is not initialized.")
         return self._database.delete_documents_partition(partition_start, partition_end)
 
     def delete_document_attributes_partition(self, partition_start: datetime, partition_end: datetime):
+        if self._database is None:
+            raise ValueError("Database is not initialized.")
         return self._database.delete_document_attributes_partition(partition_start, partition_end)
